requirements.txt
dlib
pandas
psycopg2-binary
fastapi[all]
face_recognition
yagmail[all]
python-dotenv